templating the IVF arbitrator
